Displayed list of images in run workflow modal
