+++ .env
@@ -20,3 +20,4 @@ LOG_LEVEL_CONSOLE=warn
 
 FRACTAL_RUNNER_BACKEND=local
 #WARNING_BANNER_PATH=/path/to/banner.txt
+ENABLE_INTERACTIVE_ATTRIBUTE_FILTERS=false
