+++ .env.development
@@ -19,3 +19,4 @@ LOG_LEVEL_CONSOLE=info
 
 FRACTAL_RUNNER_BACKEND=local
 #WARNING_BANNER_PATH=/path/to/banner.txt
+ENABLE_INTERACTIVE_ATTRIBUTE_FILTERS=false
