+++ docs/quickstart.md
@@ -65,6 +65,7 @@ export FRACTAL_RUNNER_BACKEND=local
 #export AUTH_COOKIE_PATH=/
 #export AUTH_COOKIE_SAME_SITE=lax
 #export PUBLIC_UPDATE_JOBS_INTERVAL=3000
+#export ENABLE_INTERACTIVE_ATTRIBUTE_FILTERS=false
 
 node build/
 ```
