+++ src/lib/components/v2/workflow/RunWorkflowModal.svelte
@@ -8,6 +8,7 @@
 	import BooleanIcon from 'fractal-components/common/BooleanIcon.svelte';
 	import Modal from '$lib/components/common/Modal.svelte';
 	import { onMount } from 'svelte';
+	import DatasetImagesTable from '../projects/datasets/DatasetImagesTable.svelte';
 
 	/** @type {import('fractal-components/types/api').DatasetV2[]} */
 	export let datasets;
@@ -21,6 +22,7 @@
 	export let onDatasetsUpdated;
 	/** @type {{[key: number]: import('$lib/types').JobStatus}} */
 	export let statuses;
+	export let attributeFiltersEnabled;
 
 	/** @type {Modal} */
 	let modal;
@@ -49,10 +51,15 @@
 		(mode === 'restart' && !replaceExistingDataset && newDatasetName === selectedDataset?.name) ||
 		(mode === 'continue' && firstTaskIndex === undefined);
 
+	/** @type {import('fractal-components/types/api').ImagePage|null} */
+	let imagePage = null;
+	/** @type {{ attributes: { [key: string]: null | string | number | boolean }, types: { [key: string]: boolean | null }} | null} */
+	let initialFilterValues = null;
+
 	/**
 	 * @param {'run'|'restart'|'continue'} action
 	 */
-	export function open(action) {
+	export async function open(action) {
 		mode = action;
 		replaceExistingDataset = true;
 		applyingWorkflow = false;
@@ -64,6 +71,7 @@
 			firstTaskIndex = getFirstTaskIndexForContinuingWorkflow(workflow.task_list, statuses);
 		}
 		lastTaskIndex = undefined;
+		await loadDatasetImages();
 		modal.show();
 	}
 
@@ -139,7 +147,9 @@
 	}
 
 	async function createNewDataset() {
-		const { zarr_dir } = /** @type {import('fractal-components/types/api').DatasetV2} */ (selectedDataset);
+		const { zarr_dir } = /** @type {import('fractal-components/types/api').DatasetV2} */ (
+			selectedDataset
+		);
 		const newDataset = await handleDatasetCreate(newDatasetName, zarr_dir);
 		onDatasetsUpdated([...datasets, newDataset], newDataset.id);
 	}
@@ -204,14 +214,18 @@
 			appliedAttributeFilters = { ...wft.input_filters.attributes };
 			appliedTypeFilters = { ...wft.input_filters.types };
 		} else {
-			const dataset = /** @type {import('fractal-components/types/api').DatasetV2} */ (selectedDataset);
+			const dataset = /** @type {import('fractal-components/types/api').DatasetV2} */ (
+				selectedDataset
+			);
 			appliedAttributeFilters = { ...dataset.filters.attributes, ...wft.input_filters.attributes };
 			appliedTypeFilters = { ...dataset.filters.types, ...wft.input_filters.types };
 		}
 	}
 
 	function computeNewDatasetName() {
-		const dataset = /** @type {import('fractal-components/types/api').DatasetV2} */ (selectedDataset);
+		const dataset = /** @type {import('fractal-components/types/api').DatasetV2} */ (
+			selectedDataset
+		);
 		newDatasetName = generateNewUniqueDatasetName(datasets, dataset.name);
 	}
 
@@ -229,12 +243,59 @@
 		}
 	}
 
+	async function loadDatasetImages() {
+		if (firstTaskIndex === undefined) {
+			return;
+		}
+
+		const task = workflow.task_list[firstTaskIndex];
+
+		const dataset = /** @type {import('fractal-components/types/api').DatasetV2} */ (
+			selectedDataset
+		);
+
+		const headers = new Headers();
+		headers.set('Content-Type', 'application/json');
+		initialFilterValues = {
+			attributes: {
+				...dataset.filters.attributes,
+				...task.input_filters.attributes
+			},
+			types: {
+				...dataset.filters.types,
+				...task.input_filters.types
+			}
+		};
+		const response = await fetch(
+			`/api/v2/project/${dataset.project_id}/dataset/${dataset.id}/images/query?page=1&page_size=10&use_dataset_filters=false`,
+			{
+				method: 'POST',
+				headers,
+				credentials: 'include',
+				body: JSON.stringify({ filters: initialFilterValues })
+			}
+		);
+		if (response.ok) {
+			imagePage = await response.json();
+		}
+	}
+
 	onMount(async () => {
 		await loadSlurmAccounts();
 	});
+
+	let opened = false;
 </script>
 
-<Modal id="runWorkflowModal" centered={true} bind:this={modal}>
+<Modal
+	id="runWorkflowModal"
+	centered={true}
+	bind:this={modal}
+	size="xl"
+	onOpen={() => (opened = true)}
+	onClose={() => (opened = false)}
+	scrollable={true}
+>
 	<svelte:fragment slot="header">
 		<h5 class="modal-title">
 			{#if mode === 'run'}
@@ -350,72 +411,108 @@
 					{/each}
 				</select>
 			</div>
-			<div class="accordion" id="accordion-workflow-advanced-options">
-				<div class="accordion-item">
-					<h2 class="accordion-header">
-						<button
-							class="accordion-button collapsed"
-							type="button"
-							data-bs-toggle="collapse"
-							data-bs-target="#collapse-workflow-advanced-options"
-							aria-expanded="false"
-							aria-controls="collapse-workflow-advanced-options"
+			{#key opened}
+				<div class="accordion" id="accordion-run-workflow">
+					<div class="accordion-item">
+						<h2 class="accordion-header">
+							<button
+								class="accordion-button collapsed"
+								type="button"
+								data-bs-toggle="collapse"
+								data-bs-target="#collapse-workflow-advanced-options"
+								aria-expanded="false"
+								aria-controls="collapse-workflow-advanced-options"
+							>
+								Advanced options
+							</button>
+						</h2>
+						<div
+							id="collapse-workflow-advanced-options"
+							class="accordion-collapse collapse"
+							data-bs-parent="#accordion-run-workflow"
 						>
-							Advanced Options
-						</button>
-					</h2>
-					<div
-						id="collapse-workflow-advanced-options"
-						class="accordion-collapse collapse"
-						data-bs-parent="#accordion-workflow-advanced-options"
-					>
-						<div class="accordion-body">
-							<div class="mb-3">
-								<label for="workerInit" class="form-label">Worker initialization (Optional)</label>
-								<textarea
-									name="workerInit"
-									id="workerInit"
-									class="form-control font-monospace"
-									rows="5"
-									disabled={checkingConfiguration}
-									bind:value={workerInitControl}
-								/>
-							</div>
-							{#if slurmAccounts.length > 0}
+							<div class="accordion-body">
 								<div class="mb-3">
-									<div class="form-check">
-										<input
-											class="form-check-input"
-											type="checkbox"
-											id="setSlurmAccount"
-											bind:checked={setSlurmAccount}
-										/>
-										<label class="form-check-label" for="setSlurmAccount">
-											Set SLURM account
-										</label>
-									</div>
+									<label for="workerInit" class="form-label">Worker initialization (Optional)</label
+									>
+									<textarea
+										name="workerInit"
+										id="workerInit"
+										class="form-control font-monospace"
+										rows="5"
+										disabled={checkingConfiguration}
+										bind:value={workerInitControl}
+									/>
 								</div>
-								{#if setSlurmAccount}
+								{#if slurmAccounts.length > 0}
 									<div class="mb-3">
-										<label for="slurmAccount" class="form-label">SLURM account</label>
-										<select
-											name="slurmAccount"
-											id="slurmAccount"
-											class="form-select"
-											disabled={checkingConfiguration}
-											bind:value={slurmAccount}
-										>
-											{#each slurmAccounts as account}
-												<option>{account}</option>
-											{/each}
-										</select>
+										<div class="form-check">
+											<input
+												class="form-check-input"
+												type="checkbox"
+												id="setSlurmAccount"
+												bind:checked={setSlurmAccount}
+											/>
+											<label class="form-check-label" for="setSlurmAccount">
+												Set SLURM account
+											</label>
+										</div>
 									</div>
+									{#if setSlurmAccount}
+										<div class="mb-3">
+											<label for="slurmAccount" class="form-label">SLURM account</label>
+											<select
+												name="slurmAccount"
+												id="slurmAccount"
+												class="form-select"
+												disabled={checkingConfiguration}
+												bind:value={slurmAccount}
+											>
+												{#each slurmAccounts as account}
+													<option>{account}</option>
+												{/each}
+											</select>
+										</div>
+									{/if}
 								{/if}
-							{/if}
+							</div>
 						</div>
 					</div>
+					{#if selectedDataset && imagePage && imagePage.images.length > 0 && firstTaskIndex !== undefined}
+						<div class="accordion-item">
+							<h2 class="accordion-header">
+								<button
+									class="accordion-button collapsed"
+									type="button"
+									data-bs-toggle="collapse"
+									data-bs-target="#collapse-workflow-image-list"
+									aria-expanded="false"
+									aria-controls="collapse-workflow-image-list"
+								>
+									Image list
+								</button>
+							</h2>
+							<div
+								id="collapse-workflow-image-list"
+								class="accordion-collapse collapse"
+								data-bs-parent="#accordion-run-workflow"
+							>
+								<div class="accordion-body">
+									<DatasetImagesTable
+										dataset={selectedDataset}
+										{imagePage}
+										{initialFilterValues}
+										{attributeFiltersEnabled}
+										useDatasetFilters={false}
+										vizarrViewerUrl={null}
+										runWorkflowModal={true}
+									/>
+								</div>
+							</div>
+						</div>
+					{/if}
 				</div>
-			</div>
+			{/key}
 			{#if checkingConfiguration}
 				<hr />
 				<h6 class="mt-3">Applied filters</h6>
